jenkins/papr: add custom jenkins imagestream
In 3.6, we don't want to use the built-in `openshift/jenkins`
imagestream because that tracks the `:latest` tag in Docker Hub rather
than a versioned stable tag. This was done starting from v3.7. For now,
let's just create our own imagestream to point to the v3.9 tag.

See link in comments for more details.